# docs/openapi.yaml
...
openapi: 3.0.4
info:
  title: dployr Base API
  version: 1.0.0
  description: |
    Your app, your server, your rules!

    ## Authentication
    Authenticate via OAuth2.0 (Google, GitHub, Microsoft) or email OTP.
    Sessions are cookie-based.

    ## Authorization
    Role-based access control:
    - **Owner**: Full control, manages admins and billing
    - **Admin**: User management, infrastructure
    - **Developer**: Deploy apps, manage integrations
    - **Viewer**: Read-only access

    ## Rate Limiting
    - Global: 100 requests/minute per user or IP on `/v1/*`
    - Strict: 10 requests/minute on sensitive endpoints
    - Returns `429` with `Retry-After` and `X-RateLimit-*` headers on breach
servers:
  - url: https://base.dployr.dev
    description: Development
  - url: https://base.dployr.io
    description: Production
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: Resource not found
            code:
              type: string
              description: |
                Machine-readable error code. Known values:
                - `request.missing_params`
                - `request.bad_request`
                - `request.invalid_otp`
                - `request.unprocessable_entity`
                - `request.too_many_requests`
                - `request.internal_server_error`
                - `validation.missing_fields`
                - `validation.invalid_format`
                - `auth.bad_session`
                - `auth.bad_token`
                - `auth.invalid_oauth_provider`
                - `auth.bad_oauth_state`
                - `permission.forbidden`
                - `permission.owner_role_required`
                - `permission.admin_role_required`
                - `permission.developer_role_required`
                - `permission.viewer_role_required`
                - `resource.missing_resource`
                - `resource.conflict`
                - `runtime.bad_webhook_signature`
                - `runtime.internal_server_error`
                - `runtime.node_not_connected`
                - `bootstrap.bootstrap_setup_failure`
                - `bootstrap.bootstrap_run_failure`
              example: resource.missing_resource
            helpLink:
              type: string
              format: uri
              example: https://status.dployr.io
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          nullable: true
          type: object
...
paths:
...
  /v1/instances/{name}/domain:
    post:
      operationId: addInstanceDomain
      tags: [Instances]
      summary: Assign free subdomain to instance
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
          description: Instance tag
      responses:
        '200':
          description: Domain assigned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          domain:
                            type: string
                            example: my-instance.dployr.io
        '400':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/domains/register:
    post:
      operationId: registerInstanceDomain
      tags: [Domains]
      summary: Register instance with bootstrap token
      description: Called by the dployr daemon on first boot using the one-time token from instance creation.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Instance registered
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          instanceId:
                            type: string
                          domain:
                            type: string
                            example: my-instance.dployr.io
                          issuer:
                            type: string
                          audience:
                            type: string
                            example: dployr-instance
        '400':
          description: Invalid, expired, or already-used token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  message: Invalid or expired token
                  code: auth.bad_token
  /v1/domains/verify:
    get:
      operationId: caddyDomainVerify
      tags: [Domains]
      summary: Caddy TLS verification endpoint
      description: |
        Called internally by Caddy to verify domain ownership before issuing TLS certificates.
        Not intended for direct API consumer use.

        Returns plain text responses:
        - `ok` — domain is active or TXT record verified (activates domain)
        - `verification pending` — TXT record not yet found (403)
        - `domain not registered` — domain unknown (404)
        - `missing domain parameter` — no `domain` query param (400)
      security: []
      parameters:
        - in: query
          name: domain
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Domain verified or already active
          content:
            text/plain:
              schema:
                type: string
                example: ok
        '400':
          description: Missing domain parameter
          content:
            text/plain:
              schema:
                type: string
                example: missing domain parameter
        '403':
          description: TXT record not yet found
          content:
            text/plain:
              schema:
                type: string
                example: verification pending
        '404':
          description: Domain not registered
          content:
            text/plain:
              schema:
                type: string
                example: domain not registered
  /v1/domains:
    get:
      operationId: listClusterDomains
      tags: [Domains]
      summary: List custom domains for a cluster
      security:
        - sessionCookie: []
      parameters:
        - in: query
          name: clusterId
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Domains
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          domains:
                            type: array
                            items:
                              type: object
                              properties:
                                id:
                                  type: string
                                domain:
                                  type: string
                                status:
                                  type: string
                                  enum: [pending, active]
                                provider:
                                  type: string
                                  nullable: true
                                createdAt:
                                  type: integer
                                  format: int64
                                activatedAt:
                                  type: integer
                                  format: int64
                                  nullable: true
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      operationId: createCustomDomain
      tags: [Domains]
      summary: Create custom domain for an instance
      description: Requires developer role on the instance's cluster.
      security:
        - sessionCookie: []
      parameters:
        - in: query
          name: instanceId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain, clusterId]
              properties:
                domain:
                  type: string
                  example: api.example.com
                clusterId:
                  type: string
                  description: cluster id
      responses:
        '200':
          description: Domain created with DNS records to configure
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          domain:
                            type: string
                          provider:
                            type: string
                            enum: [cloudflare, godaddy, digitalocean, route53, namecheap, google, unknown]
                          hasOAuth:
                            type: boolean
                          record:
                            type: object
                            description: A record pointing to instance IP
                            properties:
                              type:
                                type: string
                                enum: [A]
                              name:
                                type: string
                              value:
                                type: string
                              ttl:
                                type: integer
                          verification:
                            type: object
                            description: TXT record for ownership verification
                            properties:
                              type:
                                type: string
                                enum: [TXT]
                              name:
                                type: string
                              value:
                                type: string
                          autoSetupUrl:
                            type: string
                            nullable: true
                          manualGuideUrl:
                            type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Domain registered to another instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  message: Domain already registered to another instance
                  code: resource.conflict
  /v1/domains/{domain}:
    get:
      operationId: getDomain
      tags: [Domains]
      summary: Get domain details
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: domain
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Domain status and records (pending domains include DNS records)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          domain:
                            type: string
                          status:
                            type: string
                            enum: [pending, active]
                          clusterId:
                            type: string
                          provider:
                            type: string
                            nullable: true
                          createdAt:
                            type: integer
                            format: int64
                          activatedAt:
                            type: integer
                            format: int64
                            nullable: true
                          record:
                            type: object
                            nullable: true
                          verification:
                            type: object
                            nullable: true
                          autoSetupUrl:
                            type: string
                            nullable: true
                          manualGuideUrl:
                            type: string
                            nullable: true
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Domain not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      operationId: removeDomain
      tags: [Domains]
      summary: Remove custom domain
      description: Requires developer role on the instance's cluster.
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: domain
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Domain removed
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Domain or instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/domains/{domain}/verify:
    post:
      operationId: verifyDomain
      tags: [Domains]
      summary: Trigger client-side domain TXT verification
      description: |
        Checks the DNS TXT record and activates the domain if verified.
        Returns 204 on success, 422 if TXT not found yet.
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: domain
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Domain verified and activated
        '404':
          description: Domain or instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: TXT record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  message: TXT record not found
                  code: request.unprocessable_entity
        '500':
          description: Verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/proxy/node/routes:
    post:
      operationId: addProxyNodeRoute
      tags: [Proxy]
      summary: Add a proxy route on a node (fire-and-return)
      description: |
        Dispatches a proxy route add task to the target node and returns immediately.
        Requires developer role on the cluster.
      security:
        - sessionCookie: []
      parameters:
        - in: query
          name: clusterId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instanceName, serviceName, upstream]
              properties:
                instanceName:
                  type: string
                  description: Target instance tag/name
                serviceName:
                  type: string
                upstream:
                  type: string
                  description: Upstream address (host:port)
                domain:
                  type: string
                  nullable: true
                  description: Optional custom domain
                template:
                  type: string
                  nullable: true
                  description: Optional proxy template name
      responses:
        '202':
          description: Route add task dispatched
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          taskId:
                            type: string
                          message:
                            type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions (developer role required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Node not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/proxy/node/routes/{service}:
    delete:
      operationId: deleteProxyNodeRoute
      tags: [Proxy]
      summary: Delete a proxy route on a node (fire-and-return)
      description: |
        Dispatches a proxy route removal task to the target node and returns immediately.
        Requires developer role on the cluster.
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: service
          required: true
          schema:
            type: string
          description: Service name whose route should be removed
        - in: query
          name: instanceName
          required: true
          schema:
            type: string
          description: Target instance tag/name
        - in: query
          name: clusterId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Route removal task dispatched
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          taskId:
                            type: string
                          message:
                            type: string
        '400':
          description: Missing instanceName or service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions (developer role required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Node not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── Status (Wake-on-demand) ────────────────────────────────────────────────
...